Add the game as it was on jam
We exchanged zip files over Discord for this jam
We thought it would be faster that way
How we always use git. Merging by hand is pain...
There are 30 or so zip files by me and @sawa9889
This is our first complete game on love2d, so we probably
gonna rebuild it from scratch. But this commit is here
for history